+++ ketcher-autotests/tests/specs/Macromolecule-editor/Copy-And-Paste/flex-mode-copy-paste.spec.ts
@@ -10,10 +10,8 @@ import {
   pasteFromClipboardByKeyboard,
   moveMouseAway,
 } from '@utils';
-import {
-  pressUndoButton,
-  turnOnMacromoleculesEditor,
-} from '@tests/pages/common/TopLeftToolbar';
+import { pressUndoButton } from '@tests/pages/common/TopLeftToolbar';
+import { turnOnMacromoleculesEditor } from '@tests/pages/common/TopRightToolbar';
 import { getMonomerLocator } from '@utils/macromolecules/monomer';
 
 const startX = 300;
