+++ ketcher-autotests/tests/specs/Macromolecule-editor/Copy-And-Paste/sequence-mode-copy-paste.spec.ts
@@ -18,10 +18,8 @@ import {
   pasteFromClipboardByKeyboard,
   selectAllStructuresOnCanvas,
 } from '@utils';
-import {
-  pressUndoButton,
-  turnOnMacromoleculesEditor,
-} from '@tests/pages/common/TopLeftToolbar';
+import { pressUndoButton } from '@tests/pages/common/TopLeftToolbar';
+import { turnOnMacromoleculesEditor } from '@tests/pages/common/TopRightToolbar';
 import { waitForMonomerPreview } from '@utils/macromolecules';
 import {
   getSequenceSymbolLocator,
