+++ ketcher-autotests/tests/specs/Macromolecule-editor/Enumerations/enumerations.spec.ts
@@ -8,10 +8,8 @@ import {
   openFileAndAddToCanvasMacro,
   selectSnakeLayoutModeTool,
 } from '@utils';
-import {
-  pressUndoButton,
-  turnOnMacromoleculesEditor,
-} from '@tests/pages/common/TopLeftToolbar';
+import { pressUndoButton } from '@tests/pages/common/TopLeftToolbar';
+import { turnOnMacromoleculesEditor } from '@tests/pages/common/TopRightToolbar';
 import { getMonomerLocator } from '@utils/macromolecules/monomer';
 import {
   selectAreaSelectionTool,
