+++ ketcher-autotests/tests/specs/Macromolecule-editor/Import-Saving-Files/import-saving-idt.spec.ts
@@ -35,9 +35,12 @@ import {
   selectOpenFileTool,
   pressUndoButton,
   selectSaveTool,
+} from '@tests/pages/common/TopLeftToolbar';
+import {
   turnOnMacromoleculesEditor,
   turnOnMicromoleculesEditor,
-} from '@tests/pages/common/TopLeftToolbar';
+} from '@tests/pages/common/TopRightToolbar';
+
 import {
   closeErrorMessage,
   closeOpenStructure,
