+++ ketcher-autotests/tests/specs/Macromolecule-editor/Import-Saving-Files/import-saving-ket.spec.ts
@@ -22,9 +22,11 @@ import {
   pressUndoButton,
   selectClearCanvasTool,
   selectOpenFileTool,
+} from '@tests/pages/common/TopLeftToolbar';
+import {
   turnOnMacromoleculesEditor,
   turnOnMicromoleculesEditor,
-} from '@tests/pages/common/TopLeftToolbar';
+} from '@tests/pages/common/TopRightToolbar';
 import { selectAllStructuresOnCanvas } from '@utils/canvas';
 import { pageReload } from '@utils/common/helpers';
 import {
