+++ ketcher-autotests/tests/specs/Macromolecule-editor/Import-Saving-Files/saving-svg.spec.ts
@@ -14,8 +14,8 @@ import {
 import {
   selectClearCanvasTool,
   selectSaveTool,
-  turnOnMacromoleculesEditor,
 } from '@tests/pages/common/TopLeftToolbar';
+import { turnOnMacromoleculesEditor } from '@tests/pages/common/TopRightToolbar';
 import {
   markResetToDefaultState,
   processResetToDefaultState,
