+++ ketcher-autotests/tests/specs/Macromolecule-editor/Polymer-Bond-Tool/connection-rules-for-phosphate-monomers.spec.ts
@@ -9,10 +9,8 @@ import {
   MonomerType,
   waitForPageInit,
 } from '@utils';
-import {
-  selectClearCanvasTool,
-  turnOnMacromoleculesEditor,
-} from '@tests/pages/common/TopLeftToolbar';
+import { selectClearCanvasTool } from '@tests/pages/common/TopLeftToolbar';
+import { turnOnMacromoleculesEditor } from '@tests/pages/common/TopRightToolbar';
 import { getMonomerLocator } from '@utils/macromolecules/monomer';
 import {
   bondMonomerPointToMoleculeAtom,
