+++ ketcher-autotests/tests/specs/Macromolecule-editor/Polymer-Bond-Tool/hydrogen-bonds-for-monomers.spec.ts
@@ -26,9 +26,11 @@ import {
   pressRedoButton,
   pressUndoButton,
   selectClearCanvasTool,
+} from '@tests/pages/common/TopLeftToolbar';
+import {
   turnOnMacromoleculesEditor,
   turnOnMicromoleculesEditor,
-} from '@tests/pages/common/TopLeftToolbar';
+} from '@tests/pages/common/TopRightToolbar';
 import {
   bondSelectionTool,
   commonLeftToolbarLocators,
