+++ ketcher-autotests/tests/specs/Macromolecule-editor/RNA-Builder/rna-library.spec.ts
@@ -63,9 +63,11 @@ import {
   pressRedoButton,
   pressUndoButton,
   selectClearCanvasTool,
+} from '@tests/pages/common/TopLeftToolbar';
+import {
   turnOnMacromoleculesEditor,
   turnOnMicromoleculesEditor,
-} from '@tests/pages/common/TopLeftToolbar';
+} from '@tests/pages/common/TopRightToolbar';
 import {
   bondSelectionTool,
   selectAreaSelectionTool,
