+++ ketcher-autotests/tests/specs/Macromolecule-editor/Sequence-Mode/new-sequence-representation1-1.spec.ts
@@ -12,10 +12,8 @@ import {
   takeEditorScreenshot,
   waitForPageInit,
 } from '@utils';
-import {
-  selectClearCanvasTool,
-  turnOnMacromoleculesEditor,
-} from '@tests/pages/common/TopLeftToolbar';
+import { selectClearCanvasTool } from '@tests/pages/common/TopLeftToolbar';
+import { turnOnMacromoleculesEditor } from '@tests/pages/common/TopRightToolbar';
 import {
   getSymbolLocator,
   turnSyncEditModeOff,
