+++ ketcher-autotests/tests/specs/Macromolecule-editor/Sequence-Mode/new-sequence-representation2.spec.ts
@@ -13,10 +13,8 @@ import {
   waitForPageInit,
   waitForRender,
 } from '@utils';
-import {
-  selectClearCanvasTool,
-  turnOnMacromoleculesEditor,
-} from '@tests/pages/common/TopLeftToolbar';
+import { selectClearCanvasTool } from '@tests/pages/common/TopLeftToolbar';
+import { turnOnMacromoleculesEditor } from '@tests/pages/common/TopRightToolbar';
 import {
   getSymbolLocator,
   turnSyncEditModeOff,
