+++ ketcher-autotests/tests/specs/Macromolecule-editor/Sequence-Mode/sequence-mode-import-export.spec.ts
@@ -16,10 +16,8 @@ import {
   resetZoomLevelToDefault,
   waitForPageInit,
 } from '@utils';
-import {
-  selectClearCanvasTool,
-  turnOnMacromoleculesEditor,
-} from '@tests/pages/common/TopLeftToolbar';
+import { selectClearCanvasTool } from '@tests/pages/common/TopLeftToolbar';
+import { turnOnMacromoleculesEditor } from '@tests/pages/common/TopRightToolbar';
 import { zoomWithMouseWheel, chooseTab, Tabs } from '@utils/macromolecules';
 import { keyboardPressOnCanvas } from '@utils/keyboard/index';
