+++ ketcher-autotests/tests/specs/Macromolecule-editor/Snake-Mode/snake-mode.spec.ts
@@ -15,9 +15,11 @@ import {
 import {
   selectClearCanvasTool,
   selectSaveTool,
+} from '@tests/pages/common/TopLeftToolbar';
+import {
   turnOnMacromoleculesEditor,
   turnOnMicromoleculesEditor,
-} from '@tests/pages/common/TopLeftToolbar';
+} from '@tests/pages/common/TopRightToolbar';
 import {
   FileType,
   verifyFileExport,
