+++ ketcher-autotests/tests/specs/Macromolecule-editor/Zoom-Tool/zoom-tool.spec.ts
@@ -20,10 +20,8 @@ import {
   selectZoomReset,
   selectZoomOutTool,
 } from '@utils';
-import {
-  selectClearCanvasTool,
-  turnOnMacromoleculesEditor,
-} from '@tests/pages/common/TopLeftToolbar';
+import { selectClearCanvasTool } from '@tests/pages/common/TopLeftToolbar';
+import { turnOnMacromoleculesEditor } from '@tests/pages/common/TopRightToolbar';
 import { pageReload } from '@utils/common/helpers';
 import {
   zoomWithMouseWheel,
